pyproject.toml
[project]
name = "mediaforge"
version = "0.1.0"
description = "Unified media management application"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "python-multipart>=0.0.12",
    # Database
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    # Redis & Task Queue
    "redis>=5.2.0",
    "celery[redis]>=5.4.0",
    # HTTP Client
    "httpx>=0.28.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Configuration
    "pydantic-settings>=2.6.0",
    # Logging
    "structlog>=24.4.0",
    # File operations
    "aiofiles>=24.1.0",
    "watchdog>=6.0.0",
    # Subtitle engine
    "subliminal>=2.2.1",
    # BitTorrent (optional, requires system libtorrent)
    # "libtorrent>=2.0.0",
    # Utilities
    "jinja2>=3.1.4",
    "orjson>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
